rename file, export CloudMapping
